requirements.txt
# Core FastAPI and server dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
mangum==0.18.0

# PDF processing
pypdf==5.1.0
pdfminer.six==20231228

# NLP and text processing
spacy==3.8.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
nltk==3.9.1
locationtagger==0.0.1
urlextract==1.9.0

# Data processing
pandas==2.2.3
numpy==1.26.4

# Geolocation
geopy==2.4.1

# Web scraping (for enhanced resume parsing)
beautifulsoup4==4.12.3
requests==2.32.3

# Additional utility dependencies
python-dateutil==2.9.0
pytz==2024.2
regex==2024.11.6
tqdm==4.67.1
Pillow==11.0.0
pydantic==2.10.3
typing_extensions==4.12.2
